Fixing build etc

+++ src/createSkolem.py
@@ -25,20 +25,21 @@ THE SOFTWARE.
 import tempfile
 import os
 import numpy as np
+import subprocess
 
 def skolemfunction_preprocess(inputfile_name, Xvar,Yvar, PosUnate, NegUnate, UniqueVar = [], UniqueDef = ''):
-	
+
 	'''
 	Write the Skolem functions in the verilog format.
-	
-	Assuming y_i is positive unate: assign o_i = 1 
-	Assuming y_i is uniquely defined:  
+
+	Assuming y_i is positive unate: assign o_i = 1
+	Assuming y_i is uniquely defined:
 		assign w_i = unique_def of y_i
 		assign o_i = w_i
 
 	'''
-	
-	
+
+
 	declare = 'module SkolemFormula ('
 	declarevar = ''
 	assign = ''
@@ -67,12 +68,12 @@ def skolemfunction_preprocess(inputfile_name, Xvar,Yvar, PosUnate, NegUnate, Uni
 	with open(skolemfile_name,"w") as f:
 		f.write(skolemformula)
 	f.close()
-	
+
 
 def createSkolemfunction(inputfile_name, Xvar,Yvar):
-	
+
 	skolemformula = tempfile.gettempdir() + '/' + inputfile_name + "_skolem.v"
-	
+
 	content = ''
 	declare = "module SkolemFormula ("
 	declare_input = ""
@@ -107,9 +108,9 @@ def createSkolemfunction(inputfile_name, Xvar,Yvar):
 			line = line.replace('& o%s' %(var), "")
 			line = line.replace("o","w")
 		content += line
-	
+
 	skolemfile_name = inputfile_name + "_skolem.v"
-	
+
 	with open(skolemfile_name,"w") as f:
 		f.write(declare + declare_input + content + assign + "endmodule\n")
 	f.close()
@@ -122,10 +123,10 @@ def createSkolemfunction(inputfile_name, Xvar,Yvar):
 def createErrorFormula(Xvar, Yvar,  verilog_formula):
 
 	'''
-	
+
 	Creating E(X,Y,Y') = \varphi(X,Y) \land \lnot \varphi(X,Y') \land (Y' <-> F(X))
-	
-	inputerrorx is for X 
+
+	inputerrorx is for X
 	inputerrory is for Y
 	inputerroryp is for Y'
 
@@ -133,8 +134,8 @@ def createErrorFormula(Xvar, Yvar,  verilog_formula):
 	declare inputs for X, Y, Y'
 
 	Something like this, where |X| = 1 , and |Y| = 4
-	
-	
+
+
 	module MAIN (2, 1, 3, 4, 5, ip1, ip3, ip4, ip5, out );
 	input 2 ;
 	input 1 ;
@@ -159,7 +160,7 @@ def createErrorFormula(Xvar, Yvar,  verilog_formula):
 	'''
 
 
-	
+
 	inputformula = '('
 	inputskolem = '('
 	inputerrorx = 'module MAIN ('
@@ -175,48 +176,48 @@ def createErrorFormula(Xvar, Yvar,  verilog_formula):
 		inputskolem += "%s, " % (var)
 		inputerrorx += "%s, " % (var)
 		declarex += "input %s ;\n" % (var)
-	
+
 	for var in Yvar:
 
 		inputformula += "%s, " % (var)
 		inputerrory += "%s, " % (var)
-		declarey += "input %s ;\n" % (var) 
+		declarey += "input %s ;\n" % (var)
 		inputerroryp += "ip%s, " % (var)
 		declareyp += "input ip%s ;\n" % (var)
 		inputskolem += "ip%s, " %(var)
-		
+
 	inputformula += "out1 );\n"
 	inputformula_sk = inputskolem + "out3 );\n"
 	inputskolem += "out2 );\n"
-	
+
 	inputerrorx = inputerrorx + inputerrory + inputerroryp + "out );\n"
-	
+
 	declare = declarex + declarey + declareyp + 'output out;\n' + \
 		"wire out1;\n" + "wire out2;\n" + "wire out3;\n"
-	
+
 	formula_call = "FORMULA F1 " + inputformula
 	skolem_call = "SKOLEMFORMULA F2 " + inputskolem
 	formulask_call = "FORMULA F2 " + inputformula_sk
-	
+
 	error_content = inputerrorx + declare + \
 		formula_call + skolem_call + formulask_call
-	
+
 	error_content += "assign out = ( out1 & out2 & ~(out3) );\n" + \
 		"endmodule\n"
-	
+
 	error_content += verilog_formula
-	
-	
+
+
 	return error_content
 
 def addSkolem(error_content,inputfile_name):
 
 	skolemformula = tempfile.gettempdir() + '/' + inputfile_name + "_skolem.v"
-	
+
 	with open(skolemformula, 'r') as f:
 		skolemcontent = f.read()
 	f.close()
-	
+
 	errorformula = tempfile.gettempdir() + '/' + inputfile_name + "_errorformula.v"
 
 	with open(errorformula, "w") as f:
@@ -225,18 +226,18 @@ def addSkolem(error_content,inputfile_name):
 
 def createSkolem(candidateSkf, Xvar, Yvar, UniqueVars, UniqueDef, inputfile_name):
 
-	tempOutputFile = tempfile.gettempdir() + '/' + inputfile_name + "_skolem.v"  
-	
+	tempOutputFile = tempfile.gettempdir() + '/' + inputfile_name + "_skolem.v"
+
 	inputstr = 'module SKOLEMFORMULA ('
 	declarestr = ''
 	assignstr = ''
 	wirestr = 'wire zero;\nwire one;\n'
 	wirestr += "assign zero = 0;\nassign one = 1;\n"
 	outstr = ''
-	
+
 	itr = 1
 	wtlist = []
-	
+
 	for var in Xvar:
 
 		declarestr += "input i%s;\n" % (var)
@@ -251,7 +252,7 @@ def createSkolem(candidateSkf, Xvar, Yvar, UniqueVars, UniqueDef, inputfile_name
 		if var not in UniqueVars:
 			assignstr += 'assign w%s = (' % (var)
 			assignstr += candidateSkf[var].replace(" 1 ", " one ").replace(" 0 ", " zero ") +");\n"
-		
+
 		outstr += "(~(w%s ^ o%s)) & " % (var,var)
 		if itr % 10 == 0:
 			flag = 1
@@ -279,12 +280,12 @@ def createSkolem(candidateSkf, Xvar, Yvar, UniqueVars, UniqueDef, inputfile_name
 	inputstr += " out );\n"
 	declarestr += "output out ;\n"
 
-	
+
 	with open(tempOutputFile, "w") as f:
 		f.write(inputstr + declarestr + wirestr)
 		f.write(UniqueDef.strip("\n")+"\n")
 		f.write(assignstr + "endmodule")
-		
+
 	f.close()
 
 
@@ -304,11 +305,11 @@ def simply(inputfile_name):
 			line.replace('& o%s' %(var), "")
 			line.replace("o","w")
 		content += line
-	
+
 	with open(skolemformula,"w") as f:
 		f.write(content)
 	f.close()
-	
+
 
 
 
@@ -320,13 +321,24 @@ def verify(args, config, Xvar, Yvar, inputfile_name):
 		os.system("rm strash.txt")
 
 	file_generation_cex = config['Dependencies-Path']['file_generation_cex_path']
-	
-	cmd = "%s %s %s  > /dev/null 2>&1" % (file_generation_cex, errorformula, cexfile)
+
+	cmd = "%s %s %s" % (file_generation_cex, errorformula, cexfile)
 
 	if args.verbose >= 2:
 		print("c file generation cex --verify cmd", cmd)
 
-	os.system(cmd)
+	proc = subprocess.Popen(cmd.split(),
+				stdout=subprocess.PIPE,
+				stderr=subprocess.PIPE,
+				text=True)
+
+	stdout, stderr = proc.communicate()
+	print("STDOUT:", stdout)
+	print("STDERR:", stderr)
+	exit_code = proc.returncode  # Get exit code here
+	if exit_code != 0:
+		exit("Error in running file generation cex")
+
 	exists = os.path.isfile("strash.txt")
 	if exists:
 		os.system("rm strash.txt")
